+++ partiql-spi/src/main/kotlin/org/partiql/spi/utils/FunctionUtils.kt
@@ -4,6 +4,7 @@ import org.partiql.spi.function.FnOverload
 import org.partiql.spi.function.Function
 import org.partiql.spi.function.Parameter
 import org.partiql.spi.function.builtins.internal.PErrors
+import org.partiql.spi.internal.SqlTypeFamily
 import org.partiql.spi.types.PType
 import org.partiql.spi.utils.NumberUtils.isNumber
 import org.partiql.spi.value.Datum
@@ -91,5 +92,15 @@ internal object FunctionUtils {
         if (!value.type.isNumber()) {
             throw PErrors.unexpectedTypeException(value.type, listOf(PType.tinyint(), PType.smallint(), PType.integer(), PType.bigint(), PType.decimal(), PType.numeric(), PType.real(), PType.doublePrecision()))
         }
+    }
+
+    internal fun isDateTimeType(type: PType): Boolean {
+        return SqlTypeFamily.TIME.contains(type) ||
+            SqlTypeFamily.TIMESTAMP.contains(type) ||
+            type.code() == PType.DATE
+    }
+
+    internal fun isIntervalType(type: PType): Boolean {
+        return type.code() == PType.INTERVAL_YM || type.code() == PType.INTERVAL_DT
     }
 }